auth lib.js fix, +started extensions rework
