+++ app/sandbox/permissions/css/load.js
@@ -1,4 +1,4 @@
-const { saveReadFile } = require("../tools.js")
+const { saveReadFile } = require("../../tools.js")
 const path = require("path")
 
 function callback(data) {
