+++ app/sandbox/permissions/discord/discordRichPresence.js
