+++ app/sandbox/permissions/editor/language/registerIcons.js
@@ -1,4 +1,4 @@
-const { saveReadFile, log } = require("../tools.js")
+const { saveReadFile, log } = require("../../../tools.js")
 const path = require("path")
 
 function callback(data) {
