+++ app/sandbox/permissions/events/onInit.js
@@ -1,0 +1,9 @@
+const { ipcMain } = require("electron")
+
+function callback(data) {
+    const cb = data.selfArgs[0]
+
+    cb(data)
+}
+
+module.exports = { callback }
