+++ app/sandbox/permissions/js/clearInterval.js
