+++ app/sandbox/permissions/window/close.js
