+++ assets/css/xterm-custom.css
@@ -1,0 +1,5 @@
+.xterm-dom-renderer-owner-1 .xterm-fg-1 {
+    border-left: 2px solid #cc0000;
+    color: #ff9695!important;
+    padding-left: 10px;
+}
