+++ assets/js/extensionsHandler/events/ui/onLoadCSS.js
@@ -1,0 +1,7 @@
+export function onLoadCSSCallback({ id, content }) {
+    const style = document.createElement("style")
+    style.id = `css-${id}-${Math.floor(Math.random() * 99999)}`
+    style.textContent = content
+
+    document.head.appendChild(style)
+}
